requirements.txt
discord.py~=1.5.1
PyNaCl~=1.4.0
pandas
dnspython~=1.16.0
async-timeout~=3.0.1
aiohttp~=3.6.2